+++ assets/javascripts/kitten/components/form/password-input.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react'
+import { PasswordIcon as PasswordIconBase } from 'kitten/components/icons/password-icon'
+import { TextInput as TextInputBase } from 'kitten/components/form/text-input'
+
+const PasswordIcon = Radium(PasswordIconBase)
+const TextInput = Radium(TextInputBase)
+
+export class PasswordInput extends Component {
+  render(){
+    return (
+
+    )
+  }
+}