+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `PasswordIcon` component.
 - Feature: Add `azote` modifier on `Button` and `ButtonIcon` components.
 - Fix: Update primary colors.
 - Feature: Add new `FilterIcon` component.